+++ src/buffer_tree.cpp
@@ -92,7 +92,7 @@ BufferTree::~BufferTree() {
 	for(int l = 0; l < max_level; l++) {
 		free(flush_positions[l]);
 		free(read_buffers[l]);
-		for (uint8_t i = 0; i < B; i++) {
+		for (uint16_t i = 0; i < B; i++) {
 			free(flush_buffers[l][i]);
 		}
 		free(flush_buffers[l]);
@@ -229,7 +229,7 @@ insert_ret_t BufferTree::insert(update_t upd) {
 /*
  * Helper function which determines which child we should flush to
  */
-inline uint32_t which_child(Node key, Node min_key, Node max_key, uint8_t options) {
+inline uint32_t which_child(Node key, Node min_key, Node max_key, uint16_t options) {
 	Node total = max_key - min_key + 1;
 	double div = (double)total / options;
 	uint larger_kids = total % options;
@@ -253,7 +253,7 @@ inline uint32_t which_child(Node key, Node min_key, Node max_key, uint8_t option
  * at once otherwise the data will clash
  */
 flush_ret_t BufferTree::do_flush(char *data, uint32_t data_size, uint32_t begin, 
-	Node min_key, Node max_key, uint8_t options, uint8_t level) {
+	Node min_key, Node max_key, uint16_t options, uint8_t level) {
 	// setup
 	uint32_t full_flush = page_size - (page_size % serial_update_size);
